fix: generate the Kptfile for pipeline catalogs
so that we can upgrade them later on via `jx gitops upgrade`
